renamed http files to httprb; consistent referencing of httprb; DRY implementation
