+++ lib/new_relic/agent/instrumentation/httprb/instrumentation.rb
@@ -1,0 +1,30 @@
+# encoding: utf-8
+# This file is distributed under New Relic's license terms.
+# See https://github.com/newrelic/newrelic-ruby-agent/blob/main/LICENSE for complete details.
+
+module NewRelic::Agent::Instrumentation
+  module HTTPrb
+    def with_tracing(request)
+      wrapped_request = ::NewRelic::Agent::HTTPClients::HTTPRequest.new(request)
+  
+      begin
+        segment = NewRelic::Agent::Tracer.start_external_request_segment(
+          library: wrapped_request.type,
+          uri: wrapped_request.uri,
+          procedure: wrapped_request.method
+        )
+  
+        segment.add_request_headers wrapped_request
+  
+        response = NewRelic::Agent::Tracer.capture_segment_error(segment) { yield }
+  
+        wrapped_response = ::NewRelic::Agent::HTTPClients::HTTPResponse.new response
+        segment.process_response_headers wrapped_response
+  
+        response
+      ensure
+        segment.finish if segment
+      end
+    end
+  end
+end
