+++ lib/new_relic/agent/instrumentation/httprb/prepend.rb
@@ -1,0 +1,15 @@
+# encoding: utf-8
+# This file is distributed under New Relic's license terms.
+# See https://github.com/newrelic/newrelic-ruby-agent/blob/main/LICENSE for complete details.
+
+module NewRelic::Agent::Instrumentation
+  module HTTPrb
+    module Prepend
+      include NewRelic::Agent::Instrumentation::HTTPrb
+
+      def perform(request, options)
+        with_tracing(request) { super }        
+      end
+    end
+  end
+end
