+++ test/multiverse/suites/httprb/config/newrelic.yml
@@ -6,7 +6,7 @@ development:
   monitor_mode: true
   license_key: bootstrap_newrelic_admin_license_key_000
   instrumentation:
-    http: <%= $instrumentation_method %>
+    httprb: <%= $instrumentation_method %>
   ca_bundle_path: ../../../config/test.cert.crt
   app_name: test
   host: localhost
